First integration tests and rolled back the Caching filter
The filter seemed to fail when creating the HttpResponse (when testing was done), not the HttpRequest. For this reason, the filter is now recreated and working properly (in tests and in dev/prod environments).
All the controller methods have changed the signature to match the HttpRequest parameter instead of HttpServletRequest.
Tests pass with this change (and the application works completely).